requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.85.0
fastapi-users-db-sqlalchemy==4.0.3
frozenlist==1.3.1
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
makefun==1.15.0
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
passlib==1.7.4
pendulum==2.1.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
vk-api==11.9.9
watchfiles==0.17.0
websockets==10.3
yarl==1.8.1
zipp==3.9.0
